Add permissions. Add gist paste method

+++ src/org/mcsg/bot/skype/util/ChatManager.java
@@ -18,9 +18,6 @@ public class ChatManager {
 
 	private static ConcurrentHashMap<Chat, ArrayList<String>> chats = new ConcurrentHashMap<>();
 
-	private static int paste = 6;
-	private static int seconds = 5;
-
 	public static void printThrowable(Chat chat, Throwable t){
 		ChatManager.chat(chat, t.toString());
 		for(StackTraceElement el : t.getStackTrace())
@@ -33,7 +30,7 @@ public class ChatManager {
 				ChatManager.chat(chat, "\t" + el.toString());
 		}
 
-		for(int a  = 0; a < paste ; a++){
+		for(int a  = 0; a < Settings.Root.Bot.chat.paste ; a++){
 			chat(chat, "");
 		}
 	}
@@ -41,9 +38,9 @@ public class ChatManager {
 	private static String getUserName(User user){
 		try {return (user.getFullName().length() > 15) ? user.getFullName().substring(0,  15) : user.getFullName();} catch (Exception e) { return null; }
 	}
-	
+
 	public static void chat(Chat chat, User user, String msg){
-		chat(chat, "@"+getUserName(user)+" "+msg);
+		chat(chat, "@"+getUserName(user)+": "+msg);
 	}
 
 	public static void chat(Chat chat, String msg){
@@ -69,19 +66,20 @@ public class ChatManager {
 									sb.append(msg).append("\n");
 								}
 								sb.delete(sb.length() - 1, sb.length());
-								if(chats.remove(chat).size() > paste){
+								if(chats.remove(chat).size() > Settings.Root.Bot.chat.paste){
 									chat.send("Output: "+createPaste(sb.toString()));
 								} else {
 									try { chat.send(sb.toString()); } catch (SkypeException e) { printThrowable(chat, e); }
 								}
 							} catch (Exception e){
+								e.printStackTrace();
 								printThrowable(chat, e);
 							}
 						}
 
-						sleep(seconds * 1000);
+						sleep(Settings.Root.Bot.chat.time * 1000);
 					}catch (Exception e){
-
+						e.printStackTrace();
 					}
 
 				}
@@ -90,44 +88,37 @@ public class ChatManager {
 		}.start();
 	}
 
-	public static void setSeconds(int secondss){
-		seconds = secondss;
-	}
-
-	public static void setPaste(int pastes){
-		paste = pastes;
-	}
-
 	public static String createPaste(final String msg) throws Exception{
-		String paste = "files/paste"+System.currentTimeMillis();
-		File file1 = new File(paste);
-		{
-			PrintWriter pw = new PrintWriter(new FileWriter(file1));
-			pw.println(msg);
-			pw.flush();
-			pw.close();
-		}
-		String script = "files/script" + System.currentTimeMillis();
-		File file2 = new File(script);
-		{
-			PrintWriter pw1 = new PrintWriter(new FileWriter(file2));
-			pw1.println("cat "+paste+" | pastebinit -a \"MC-SG.BOT Output\"");
-			pw1.flush();
-			pw1.close();
-		}
+		if(Settings.Root.Bot.chat.pastemethod.equals("gist")){
+			return GistPaster.paste(msg);
+		} else {
+			String paste = "files/paste"+System.currentTimeMillis();
 
-		Process proc = Runtime.getRuntime().exec("bash "+script);
-		String line = "";
+			File file1 = new File(paste);
+			FileUtils.writeFile(file1, msg);
+			
+			String script = "files/script" + System.currentTimeMillis();
+			File file2 = new File(script);
+			{
+				PrintWriter pw1 = new PrintWriter(new FileWriter(file2));
+				pw1.println("cat "+paste+" | pastebinit -a \"MC-SG.BOT Output\"");
+				pw1.flush();
+				pw1.close();
+			}
 
-		file2.deleteOnExit();
-		file2.deleteOnExit();
+			Process proc = Runtime.getRuntime().exec("bash "+script);
+			String line = "";
 
-		BufferedReader in = new BufferedReader(
-				new InputStreamReader(proc.getInputStream()));
-		while ((line = in.readLine()) != null) {
+			file1.deleteOnExit();
+			file2.deleteOnExit();
+
+			BufferedReader in = new BufferedReader(
+					new InputStreamReader(proc.getInputStream()));
+			while ((line = in.readLine()) != null) {
+				return line;
+			}
 			return line;
 		}
-		return line;
 	}
